FormCore v0.3.0

+++ Backend/Models/Section.cs
@@ -9,7 +9,7 @@ namespace FormCore
     public int Id { get; set; }
     public int FormId { get; set; }
     public string Title { get; set; }
-    public int Position { get; set; }
+    public double Position { get; set; }
 
     public virtual Form Form { get; set; }
     public virtual ICollection<Field> Fields { get; set; }
